requirements.txt
fastapi==0.115.6
alt-profanity-check==1.5.2
uvicorn==0.34.0
python-dotenv
sqlalchemy 
psycopg2-binary